Resolve wait for workers TODO
Previously, when switching between unicorn master processes there would
be a 2 minute wait for the new worker processes to be created. This
replaces that wait with a loop that checks whether the new process has
the same number of workers as the old process. If it does it stops
waiting and starts shutting down the old process.

Should the same number of workers not return on the new process the same
2 minute will expire and the new master process would be shut down as
before. This could happen in valid circumstances, such as someone
reducing the number of workers in a deploy, or if someone had manually
upped the old process workers with TTOU signals.

This has been done to reduce the amount of time that two applications
are running simultaneously, this prolonged period can cause problems as
any monitoring to check whether a deploy is ok has to wait for this to
change - or risks checking with a 50/50 chance of which version of an
app to be served.